ASWCrypter Update
I love this crypter by the way. I've used it quite a bit. One thing that I'd like to see is the option to not stop the postgresql and metasploit services when exiting the script. I'll usually pull up msf and then use ASW to crypt something and still want to use metasploit after I'm done with it. I didn't get a chance to test it, but I'll do that asap.