ASWCrypter Update
I love your crypter by the way. I use it a lot. Something I'd like to see is the option to not stop the postgresql and msf services when ASW exits. I sometimes use ASWCrypter while in a metasploit session and don't necessarily want to exit msf and postgres. Keep up the good work.